+++ apps/restconf/restconf_main_native.c
@@ -1405,7 +1405,7 @@ main(int    argc,
      * @see clicon_hello_req
      */
     clicon_data_set(h, "session-transport", "cl:restconf");
-
+    clixon_log(h, LOG_NOTICE, "%s: %u Started", __PROGRAM__, getpid());
     /* Main event loop */
     if (clixon_event_loop(h) < 0)
         goto done;
@@ -1416,6 +1416,8 @@ main(int    argc,
     if (xrestconf)
         xml_free(xrestconf);
     if (h){
+        clixon_log_init(h, __PROGRAM__, LOG_INFO, 0); /* Log on syslog no stderr */
+        clixon_log(h, LOG_NOTICE, "%s: %u Terminated", __PROGRAM__, getpid());
         restconf_native_terminate(h);
         restconf_terminate(h);
     }